requirements.txt
absl-py==0.2.1
appnope==0.1.0
astor==0.6.2
biopython==1.71
bleach==1.5.0
certifi==2018.4.16
chardet==3.0.4
cycler==0.10.0
decorator==4.1.2
dj-database-url==0.5.0
Django==2.0.6
django-bootstrap-toolkit==2.15.0
django-registration-redux==2.4
entrypoints==0.2.3
enum34==1.1.6
future==0.16.0
gast==0.2.0
grpcio==1.12.0
gunicorn==19.8.1
h5py==2.7.1
html5lib==0.9999999
hyperas==0.4
hyperopt==0.1
idna==2.6
ipykernel==4.6.1
ipython==5.0.0
ipython-genutils==0.2.0
ipywidgets==7.0.5
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.0
jupyter-core==4.4.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.4
jupyter-nbextensions-configurator==0.4.0
Keras==2.1.2
lxml==4.2.1
Markdown==2.6.9
MarkupSafe==1.0
matplotlib==2.1.0
missingno==0.4.0
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.0
nose==1.3.7
notebook==5.2.2
numpy==1.13.3
olefile==0.44
opencv-python==3.4.1.15
pandas==0.21.0
pandas-datareader==0.6.0
pandocfilters==1.4.2
patsy==0.5.0
pexpect==4.3.0
pickleshare==0.7.4
Pillow==4.3.0
prompt-toolkit==1.0.15
protobuf==3.5.0.post1
psycopg2==2.7.5
ptyprocess==0.5.2
py==1.5.2
Pygments==2.2.0
pymongo==3.6.0
pyparsing==2.2.0
pytest==3.2.5
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
pyzmq==16.0.3
qtconsole==4.3.1
requests==2.18.4
requests-file==1.4.3
requests-ftp==0.3.1
scikit-learn==0.19.1
scipy==1.0.0
seaborn==0.8.1
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
tensorboard==1.8.0
tensorflow==1.8.0
tensorflow-tensorboard==0.4.0rc3
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
Theano==1.0.1
tornado==4.5.2
traitlets==4.3.2
urllib3==1.22
virtualenv==16.0.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
whitenoise==3.3.1
widgetsnbextension==3.0.8
wrapt==1.10.11
xgboost==0.71
